Removed old files, moved configurable paths to text file, and improved readme
There was an old OWLSIZVerbaliser that was retained, for some reason. I've removed it.
I've also added a guide on how to resolve issues pertaining to using the error detection module.
Removed all the file paths freom code to the config.properties file to ensure they can be changed easily.